Move 'update_auctionPeriod' func. to utils module
