+++ openprocurement/auction/tests/functional/main.py
@@ -4,38 +4,15 @@ from gevent import monkey
 monkey.patch_all()
 
 import os.path
-import datetime
-import json
 import sys
 import argparse
-import contextlib
-import tempfile
-from dateutil.tz import tzlocal
-from pkg_resources import iter_entry_points
 from gevent.subprocess import check_output, sleep
+from ..utils import update_auctionPeriod
 from robot import run_cli
 
 
 PWD = os.path.dirname(os.path.realpath(__file__))
 CWD = os.getcwd()
-
-
-@contextlib.contextmanager
-def update_auctionPeriod(path, auction_type):
-    with open(path) as file:
-        data = json.loads(file.read())
-    new_start_time = (datetime.datetime.now(tzlocal()) + datetime.timedelta(seconds=120)).isoformat()
-    if auction_type == 'simple':
-        data['data']['auctionPeriod']['startDate'] = new_start_time
-    elif auction_type == 'multilot':
-        for lot in data['data']['lots']:
-            lot['auctionPeriod']['startDate'] = new_start_time
-
-    with tempfile.NamedTemporaryFile(delete=False) as auction_file:
-        json.dump(data, auction_file)
-        auction_file.seek(0)
-    yield auction_file.name
-    auction_file.close() 
 
 
 def run_simple(tender_file_path, auction_id):
