+++ openprocurement/auction/tests/utils.py
@@ -1,0 +1,23 @@
+import datetime
+import json
+import contextlib
+import tempfile
+from dateutil.tz import tzlocal
+
+
+@contextlib.contextmanager
+def update_auctionPeriod(path, auction_type):
+    with open(path) as file:
+        data = json.loads(file.read())
+    new_start_time = (datetime.datetime.now(tzlocal()) + datetime.timedelta(seconds=120)).isoformat()
+    if auction_type == 'simple':
+        data['data']['auctionPeriod']['startDate'] = new_start_time
+    elif auction_type == 'multilot':
+        for lot in data['data']['lots']:
+            lot['auctionPeriod']['startDate'] = new_start_time
+
+    with tempfile.NamedTemporaryFile(delete=False) as auction_file:
+        json.dump(data, auction_file)
+        auction_file.seek(0)
+    yield auction_file.name
+    auction_file.close()
